telephony: Hack GSM and LTE signal strength
makeSignalStrengthFromRilParcel() is used for both solicited and
unsolicited requests, so we can hack it to send better signal strength
values to the modem backend. This significantly helps in areas where
GSM or LTE signal is poor but usable, in which case the modem would often
disconnect completely or fall back to a lower network mode (resulting in
poor UX).

Revised to fix the invalid GSM signal strength values (over 31 ASU).

Change-Id: Idf86ae6e5f091ebd334542103c5133a404e757ef
Signed-off-by: mydongistiny <jaysonedson@gmail.com>